finish sponsor module

+++ imooc-ad-service/ad-sponsor/src/main/java/com/imooc/ad/entity/Creative.java
@@ -37,7 +37,7 @@ public class Creative {
     private Long size;
 
     @Column(name = "duration", nullable = false)
-    private Long duration;
+    private Integer duration;
 
     @Column(name = "audit_status", nullable = false)
     private Integer auditStatus;
